Create 150. Evaluate Reverse Polish notation
We use stack here to solve this postfix expression. 

Initially, we declare a stack and traverse through the array iteratively. There are two cases:
1. If the string is an integer which we check by using isdigit function in STL of C++, then we push it into the stack.
(We consider negative numbers by checking token.size()>1)

2. If the string is an operand, we pop the top two elements and perform the operation and push it back onto the stack. 

We continue these steps as we traverse through the array and return the top of the stack as the answer.